+++ woo-payzen-payment/includes/Lyranetwork/Payzen/Sdk/Rest/Api.php
@@ -116,6 +116,15 @@ class Api
         return $this->fallbackPost($target, $data);
     }
 
+    protected function curlClose($curl)
+    {
+        if (PHP_VERSION_ID >= 80000) {
+            unset($curl);
+        } else {
+            curl_close($curl);
+        }
+    }
+
     /**
      * @param string $target
      * @param mixed $data
@@ -149,7 +158,7 @@ class Api
         if (! in_array($info['http_code'], array(200, 401), true)) {
             $error = curl_error($curl);
             $errno = curl_errno($curl);
-            curl_close($curl);
+            $this->curlClose($curl);
 
             $msg = "Call to URL $url failed with unexpected status: {$info['http_code']}";
 
@@ -170,7 +179,7 @@ class Api
         if (! is_array($response)) {
             $error = curl_error($curl);
             $errno = curl_errno($curl);
-            curl_close($curl);
+            $this->curlClose($curl);
 
             $msg = "Call to URL $url returned an unexpected response, raw response: $raw_response";
 
@@ -183,7 +192,7 @@ class Api
             throw new \Exception($msg, '-1');
         }
 
-        curl_close($curl);
+        $this->curlClose($curl);
 
         return $response;
     }